project: Fix bugs

+++ src/VMConfig.h
@@ -12,7 +12,7 @@ namespace Pip2 {
         PoolItems pool_items_;
 
     public:
-        explicit VMConfig(std::uint8_t *memory_base, std::size_t memory_size, const std::uint32_t *pool_items_base, std::size_t pool_item_count)
+        explicit VMConfig(std::uint8_t *memory_base, std::size_t memory_size, const std::uint64_t *pool_items_base, std::size_t pool_item_count)
             : memory_base_(memory_base), memory_size_(memory_size), pool_items_(pool_items_base, pool_item_count) {
         }
